[MultiFunc] Revise file structure and extract CallGraph utilities.
